+++ src/components.d.ts
@@ -6,8 +6,58 @@
  */
 import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
 import { TecnologiaTheme } from "./components/interfaces";
+import { TecStatus } from "./models/status.model";
+import { TecButtonColor, TecButtonIconMode, TecButtonMode, TecButtonSize } from "./components/tec-button/tec-button.model";
 export namespace Components {
     interface MyComponent {
+        "theme": TecnologiaTheme;
+    }
+    interface TecButton {
+        /**
+          * Optional ID to be attached on button
+         */
+        "buttonId": string;
+        /**
+          * Colors of button (like gradient)
+         */
+        "color": TecButtonColor;
+        /**
+          * Boolean to indicate if button is disabled
+         */
+        "disabled": boolean;
+        /**
+          * If `true` button use `width: 100%`
+         */
+        "fullWidth": boolean;
+        /**
+          * Icon class from FontAwesome 5 Free Allows to use: brands, regular, solid Example: 'far fa-paper-plane'
+         */
+        "icon": string;
+        /**
+          * Position of icon
+         */
+        "iconMode": TecButtonIconMode;
+        /**
+          * Text to show inside button
+         */
+        "label": string;
+        "loading": boolean;
+        /**
+          * Mode of button (like square or rounded)
+         */
+        "mode": TecButtonMode;
+        /**
+          * If `true` button removes label
+         */
+        "onlyIcon": boolean;
+        /**
+          * Size of button
+         */
+        "size": TecButtonSize;
+        /**
+          * The status of button (color)
+         */
+        "status": TecStatus;
         "theme": TecnologiaTheme;
     }
     interface TecProductHeader {
@@ -36,6 +86,12 @@ declare global {
         prototype: HTMLMyComponentElement;
         new (): HTMLMyComponentElement;
     };
+    interface HTMLTecButtonElement extends Components.TecButton, HTMLStencilElement {
+    }
+    var HTMLTecButtonElement: {
+        prototype: HTMLTecButtonElement;
+        new (): HTMLTecButtonElement;
+    };
     interface HTMLTecProductHeaderElement extends Components.TecProductHeader, HTMLStencilElement {
     }
     var HTMLTecProductHeaderElement: {
@@ -44,11 +100,64 @@ declare global {
     };
     interface HTMLElementTagNameMap {
         "my-component": HTMLMyComponentElement;
+        "tec-button": HTMLTecButtonElement;
         "tec-product-header": HTMLTecProductHeaderElement;
     }
 }
 declare namespace LocalJSX {
     interface MyComponent {
+        "theme"?: TecnologiaTheme;
+    }
+    interface TecButton {
+        /**
+          * Optional ID to be attached on button
+         */
+        "buttonId"?: string;
+        /**
+          * Colors of button (like gradient)
+         */
+        "color"?: TecButtonColor;
+        /**
+          * Boolean to indicate if button is disabled
+         */
+        "disabled"?: boolean;
+        /**
+          * If `true` button use `width: 100%`
+         */
+        "fullWidth"?: boolean;
+        /**
+          * Icon class from FontAwesome 5 Free Allows to use: brands, regular, solid Example: 'far fa-paper-plane'
+         */
+        "icon"?: string;
+        /**
+          * Position of icon
+         */
+        "iconMode"?: TecButtonIconMode;
+        /**
+          * Text to show inside button
+         */
+        "label": string;
+        "loading"?: boolean;
+        /**
+          * Mode of button (like square or rounded)
+         */
+        "mode"?: TecButtonMode;
+        /**
+          * Emitted when button is clicked Captured by onClick listener. > Note: if button was disabled event can't be dispatch
+         */
+        "onClicked"?: (event: CustomEvent<UIEvent>) => void;
+        /**
+          * If `true` button removes label
+         */
+        "onlyIcon"?: boolean;
+        /**
+          * Size of button
+         */
+        "size"?: TecButtonSize;
+        /**
+          * The status of button (color)
+         */
+        "status"?: TecStatus;
         "theme"?: TecnologiaTheme;
     }
     interface TecProductHeader {
@@ -76,6 +185,7 @@ declare namespace LocalJSX {
     }
     interface IntrinsicElements {
         "my-component": MyComponent;
+        "tec-button": TecButton;
         "tec-product-header": TecProductHeader;
     }
 }
@@ -84,6 +194,7 @@ declare module "@stencil/core" {
     export namespace JSX {
         interface IntrinsicElements {
             "my-component": LocalJSX.MyComponent & JSXBase.HTMLAttributes<HTMLMyComponentElement>;
+            "tec-button": LocalJSX.TecButton & JSXBase.HTMLAttributes<HTMLTecButtonElement>;
             "tec-product-header": LocalJSX.TecProductHeader & JSXBase.HTMLAttributes<HTMLTecProductHeaderElement>;
         }
     }